move to right class
