+++ api-tester/src/defaults/java/com/revenuecat/apitester/java/PurchasesAPI.java
@@ -7,16 +7,22 @@ import androidx.annotation.NonNull;
 import com.revenuecat.purchases.CacheFetchPolicy;
 import com.revenuecat.purchases.CustomerInfo;
 import com.revenuecat.purchases.EntitlementVerificationMode;
+import com.revenuecat.purchases.LogLevel;
 import com.revenuecat.purchases.Purchases;
 import com.revenuecat.purchases.PurchasesConfiguration;
 import com.revenuecat.purchases.PurchasesError;
 import com.revenuecat.purchases.Store;
+import com.revenuecat.purchases.amazon.AmazonConfiguration;
 import com.revenuecat.purchases.interfaces.LogInCallback;
 import com.revenuecat.purchases.interfaces.ReceiveCustomerInfoCallback;
 import com.revenuecat.purchases.interfaces.SyncPurchasesCallback;
+import com.revenuecat.purchases.models.BillingFeature;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
 
@@ -104,5 +110,18 @@ final class PurchasesAPI {
         Purchases.configure(purchasesConfiguration);
 
         final boolean debugLogs = Purchases.getDebugLogsEnabled();
+    }
+
+    static void checkAmazonConfiguration(final Context context,
+                                         final ExecutorService executorService) {
+        PurchasesConfiguration amazonConfiguration = new AmazonConfiguration.Builder(context, "")
+                .appUserID("")
+                .observerMode(true)
+                .observerMode(false)
+                .service(executorService)
+                .diagnosticsEnabled(true)
+                .entitlementVerificationMode(EntitlementVerificationMode.INFORMATIONAL)
+                .showInAppMessagesAutomatically(true)
+                .build();
     }
 }
